Time: 37 ms (17.48%), Space: 45.3 MB (13.9%) - LeetHub

+++ 1460-number-of-substrings-containing-all-three-characters/1460-number-of-substrings-containing-all-three-characters.java
@@ -2,30 +2,30 @@ class Solution {
     public int numberOfSubstrings(String s) {
         Map<Character,Integer> map = new HashMap<>();
 
-        long l=0; 
-        long r=0; 
+        int l=0; 
+        int r=0; 
         long count =0;
 
         while(r < s.length()){
             int f =1;
-            if(map.containsKey(s.charAt((int)r))){
-                    f =  map.get(s.charAt((int)r)) +1;
+            if(map.containsKey(s.charAt(r))){
+                    f =  map.get(s.charAt(r)) +1;
             }
-            map.put(s.charAt((int)r),f);
+            map.put(s.charAt(r),f);
 
             while(map.size() >= 3){
-               int val = map.get(s.charAt((int)l))-1;
+               int val = map.get(s.charAt(l))-1;
                if(val == 0){
-                map.remove(s.charAt((int)l));
+                map.remove(s.charAt(l));
                }
                else{
-                map.put(s.charAt((int)l),val);
+                map.put(s.charAt(l),val);
                }
                l++;
             
 
             }
-            count += (r-l+1);
+            count += (long)(r-l+1);
             r++;
         }
